Refactored design-fields into individual component files
Split monolithic design-fields.tsx into a design-fields/ directory with
one file per component for better maintainability and code navigation.
