+++ apps/admin-x-settings/src/components/settings/email-design/design-fields/background-color-field.tsx
@@ -1,0 +1,13 @@
+import ColorPickerField from '../color-picker-field';
+import {useEmailDesign} from '../email-design-context';
+
+export const BackgroundColorField = () => {
+    const {settings, onSettingsChange} = useEmailDesign();
+    return (
+        <ColorPickerField
+            title="Background color"
+            value={settings.background_color}
+            onChange={color => onSettingsChange({background_color: color})}
+        />
+    );
+};
