+++ apps/admin-x-settings/src/components/settings/email-design/design-fields/button-color-field.tsx
@@ -1,0 +1,13 @@
+import ColorPickerField from '../color-picker-field';
+import {useEmailDesign} from '../email-design-context';
+
+export const ButtonColorField = () => {
+    const {settings, onSettingsChange, accentColor} = useEmailDesign();
+    return (
+        <ColorPickerField
+            title="Button color"
+            value={settings.button_color || accentColor}
+            onChange={color => onSettingsChange({button_color: color})}
+        />
+    );
+};
