+++ apps/admin-x-settings/src/components/settings/email-design/design-fields/header-background-field.tsx
@@ -1,0 +1,13 @@
+import ColorPickerField from '../color-picker-field';
+import {useEmailDesign} from '../email-design-context';
+
+export const HeaderBackgroundField = () => {
+    const {settings, onSettingsChange} = useEmailDesign();
+    return (
+        <ColorPickerField
+            title="Header background color"
+            value={settings.header_background_color}
+            onChange={color => onSettingsChange({header_background_color: color})}
+        />
+    );
+};
