+++ apps/admin-x-settings/src/components/settings/email-design/design-fields/index.ts
@@ -1,0 +1,15 @@
+export {BackgroundColorField} from './background-color-field';
+export {BodyFontField} from './body-font-field';
+export {ButtonColorField} from './button-color-field';
+export {ButtonCornersField} from './button-corners-field';
+export {ButtonStyleField} from './button-style-field';
+export {DividerColorField} from './divider-color-field';
+export {HeaderBackgroundField} from './header-background-field';
+export {HeadingFontField} from './heading-font-field';
+export {HeadingWeightField} from './heading-weight-field';
+export {ImageCornersField} from './image-corners-field';
+export {LinkColorField} from './link-color-field';
+export {LinkStyleField} from './link-style-field';
+export {PostTitleColorField} from './post-title-color-field';
+export {SectionTitleColorField} from './section-title-color-field';
+export {TitleAlignmentField} from './title-alignment-field';
