+++ apps/admin-x-settings/src/components/settings/email-design/design-fields/link-style-field.tsx
@@ -1,0 +1,17 @@
+import {Bold, Italic, Underline} from 'lucide-react';
+import {ToggleGroup, ToggleGroupItem} from '@tryghost/shade';
+import {useEmailDesign} from '../email-design-context';
+
+export const LinkStyleField = () => {
+    const {settings, onSettingsChange} = useEmailDesign();
+    return (
+        <div className="flex items-center justify-between">
+            <span className="text-sm">Link style</span>
+            <ToggleGroup type="single" value={settings.link_style || 'underline'} onValueChange={(value: string) => value && onSettingsChange({link_style: value})}>
+                <ToggleGroupItem aria-label="Underline" value="underline"><Underline className="size-4" /></ToggleGroupItem>
+                <ToggleGroupItem aria-label="Italic" value="regular"><Italic className="size-4" /></ToggleGroupItem>
+                <ToggleGroupItem aria-label="Bold" value="bold"><Bold className="size-4" /></ToggleGroupItem>
+            </ToggleGroup>
+        </div>
+    );
+};
